bring test coverage above 90% for Card and the Game classes
