+++ src/main/java/controllers/ApplicationController.java
@@ -17,6 +17,8 @@
 package controllers;
 
 import models.Game;
+import models.EnglishGame;
+import models.SpanishGame;
 import ninja.Context;
 import ninja.Result;
 import ninja.Results;
@@ -36,8 +38,18 @@ public class ApplicationController {
         return Results.html().template("views/AcesUp/AcesUp.flt.html");
     }
     
-    public Result gameGet(){
-        Game g = new Game();
+    public Result englishGameGet(){
+        EnglishGame g = new EnglishGame();
+        g.buildDeck();
+        g.shuffle();
+        g.dealFour();
+        g.error = false;
+
+        return Results.json().render(g);
+    }
+
+    public Result spanishGameGet(){
+        SpanishGame g = new SpanishGame();
         g.buildDeck();
         g.shuffle();
         g.dealFour();
